perf: optimizing the code
